requirements.txt
django
gunicorn
django_bootstrap5
libsass
django-compressor
django-sass-processor
requests
psycopg2-binary==2.9.3
djangorestframework
django-cors-headers